Corrects the errors with Symfony 3
It looks like in previous version of symdony it needs to know about
services that Symfony 4 just ignore.

So I change the configuration of my testing kernel to register those
services.
They were used for the tests of the command, I took the configuration
from there and put it back directly inside the kernel.

Plus, add tests/cache into the .gitignore
In Symfony 3 that's where it saves the cache, instead of /var/cache in
Symfony 4

And I also ignore this new directory for phpcs