Adapts to the changes of elythyr/prooph-fixtures
Adds a ProjectionsNamesProvider to provides all the projections names
to the new projections cleaner.

Adds a ProjectionFactory to help to create the projections.

Adds a PdoCleaningProjectionStrategy which is an implementation of the
CleaningProjectionStrategy for the projections cleaner.
It allows to handle the cleaning of projections which are running or not
in the correct way.

Reorganizes the tests into a fake bundle.
It's a bit clearer and closer to the real use.