Updates PdoCleaningProjectionStrategy
Tests that the exceptions are handled correctly to corresponds to the
types of exceptions delcared by the CleaningProjectionStrategy
interface.

I didn't test it first because it seems trivial...
At the end I fogot the use for Throwable so it didn't work! :D
Plus it allows me to get to 100% code coverage.